requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
werkzeug==2.0.1
ultralytics==8.0.196
opencv-python==4.8.0.74
numpy==1.24.3
pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
pymysql==1.1.0
cryptography==41.0.1 